Update `isRealmEnabled` sql
The previous SQL statement answered the question, Is there a $realm that has
been installed. This updated version answers the question, Is there an acl that
grants access to this realm. This corresponds w/ having an entry in
`roles.json#<role>#query_descripters` w/ a `realm` property = $realm, that is
both enabled and visible.