Adding additional Tests
Added tests for `controllers/user_interface.php/operation=get_menus` and for
`Acls::getDisabledMenus`. `get_menus` is the endpoint that provides the list of
of query descriptors visible to the user. While `getDisabledMenus` provides the
list of query descriptors that are disabled. This should provide us a with a set
of happy path tests for these two sets of functionality that can later be
expanded upon as needed.