Fix Query Descriptor Visibility (#911)
* Updated to allow query descriptors to control visibility

** acl-config **
Previously we only allowed statistics to determine if an `acl_group_by` record
was visible or not. But our config files seem to indicate that certain query
descriptors should be able to control their ability to be shown as well. To
support both of these use cases the `processQueryDescriptors` function in
`acl-config` has been updated to take into account the `hide` property defined
at the query descriptor level and the `visible` property defined at the
statistic level. The logic used to determine the final visibility of an
`acl_group_by` record is as follows.

- By default all entries are shown
- If `visible` is defined then we use it's value.
- Else, if `hide` is defined then we use it's value.

Also, there was a minor refactor with the removal of the `$enabled` variable.
The `$disable` variable that it was based on was never going to be null, so it's
use could be replaced w/ just `!$disable`.

** Acls **
There was some minor cleanup done because it was determined that we always want
to have the `visible` column included, not just when a `$statisticName` is
provided.

This allowed us to move the `$descripter->setShowMenu(...)` from out of the
`if(isset($statisticName)) { ... }` guard so that all query descriptors
generated by this function will have their `showMenu` property populated.

These two sets of changes work together to allow us to control our UI state as
we expect to.

* Fixing Tests

Fixing issues introduced by #897. Mostly it was just updating hard coded paths,
but there are a few places where the format of the expected data json files are
important ( ETLv2 Directory Scanner Tests ). This required re-formatting of the
affected files per their use case. The

* Adding a step to account for the User Profile Prompting

The User Profile pops up to prompt the user to double check their email address.
This means that the ext-mask-msg will be present until the user clicks close,
thus causing a problem for the `Logout` step.

* Style Fixes

* Adding tests for `get_menus` endpoint

This is just the initial bootstrap commit for the new `get_menus` endpoint
integration tests.

* Adding additional Tests

Added tests for `controllers/user_interface.php/operation=get_menus` and for
`Acls::getDisabledMenus`. `get_menus` is the endpoint that provides the list of
of query descriptors visible to the user. While `getDisabledMenus` provides the
list of query descriptors that are disabled. This should provide us a with a set
of happy path tests for these two sets of functionality that can later be
expanded upon as needed.

* Updated artifacts to be posix compliant

aka. added newlines