Re-Adding the `getResources` function
After much research and discussion w/ @jpwhite4, this function is being re-added w/ some
additional documentation && an authorization check to ensure that it is only
called for Center [Director|Staff] users. In addition, the REST endpoint that
uses this function ( in the `appkernel` module )is also being updated to require
users who request it to be Center [Director|Staff] as well.