Novice User Interface - Center Report Card: Support (#943)
* Initial Prototype of Novice User Portal

* Initial Commit

* Removing unrelated files

These files belong in the `appkernels` module and have been moved there.

* Linting Cleanup

Just cleaning things up to appease the linter.

* Removing unused file

This should not have been committed.

* Removing unused code

This change is included in the appkernels module branch `novice_appkernels`.

* Unneeded code

The inclusion of these files should be via AppKernels assets.d file.

* Update classes/Rest/Controllers/SummaryControllerProvider.php

Updates per @jpwhite4 code review comments

Co-Authored-By: Joe White <jpwhite4@buffalo.edu>

* Adding a new `includePub` function parameter

This is so that we can control when `$user->getAllRoles()` includes the public
acl. The default value for all arguments has been set to `true` as that was the
behavior pre-change. This will be used in the `AppKernels` module to help filter
appropriately for the Center Report Card Portlet.

* Removing unused function

We're no longer using this function so it is being removed.

* Added missing function parameter

Was missing a parameter from the `checkDataAccess` function call.

* Re-Adding the `getResources` function

After much research and discussion w/ @jpwhite4, this function is being re-added w/ some
additional documentation && an authorization check to ensure that it is only
called for Center [Director|Staff] users. In addition, the REST endpoint that
uses this function ( in the `appkernel` module )is also being updated to require
users who request it to be Center [Director|Staff] as well.

* Updated `getResources` sql

Just needed to update the sql so that it returns data in the same format as
`MetricExplorer::getDimensionValues` as that is what `AppKernels/PerformanceMap`
is currently setup to process.