Fixed detecting when an SSO User is being edited (#954)
* Fixed detecting when an SSO User is being edited

Due to a previous change, the initialization of `cached_user_type` was moved to
inside the `else` clause. Which makes it a little hard when it's being used in
the `if` clause. I've updated the code so that it correctly detects if an SSO

* Updates per @jpwhite4 convo

So the gist of the changes in this commit are:
  - remove the `cached_user_type` variable as it's no longer needed.
  - instead of relying on `cached_user_type` to persist data, just set
    `cmbUserType.value` like we do with every other piece of user data.
  - Make sure to blank the `cmbUserType` value on load like we do for the email
    text box.

* Removing unused variable

* Updates per code review @jpwhite4

* Removing unused code.

* Fix problem with changing user type causing infinite save loop

This was causing a problem that caused the user to get into an infinite change
loop unless they selected to save the user changes. Changes that didn't really
exist.

The problem was the listeners attached to the User Type `change|select` events
`comboChangeHandler`. This function would actually cause the `Manual Override`
checkbox to have its value updated, thus making it look like the user settings
had been changed when they had not been.

Now we just call `updateSaveIndicator` which does not modify any field values
and only controls showing / hiding the save indicator.