Novice User Interface - Summary Statistics Portlet (#930)
* Initial Prototype of Novice User Portal

* Added the SummaryStatisticsPortlet

This adds a new portlet, `SummaryStatisticsPortlet`, to OpenXDMoD. It provides
the same functionality as the summary statistics bar at the top of the old
Summary page ( via `summary3.php` ).

- The format of the statistics have been refactored out of `summary3.php` and
into a configuration file that can be modified:
`configuration/roles.d/summary_statistics.json`.
- The portlet is made visible to users via: `configuration/roles.d/summary.json`

* Added new Integration Tests and artifacts

Added a new integration test to exercise the `summary/statistics` REST endpoint
and the associated artifacts.

* Really adding the artifacts

* Adding last artifact

Missed one.

* Updated Tests

The changes to `SummaryControllerProviderTest.php` were needed to account for
validating `float` data values differently then other values.

Also updated the expected error message for start_date after end_date

* Updates per @jpwhite4 code review comments

- Removed `DeveloperPortlet.js` as it shouldn't have been committed in the first
  place.
- Moved `SummaryControllerProviderTest` & it's related test artifacts to the
  `rest` sub dir as opposed to `controllers`.
- Updated comment @ the end of `SummaryStatisticsPortlet.js` to be correct.
- re: returning 400's vs 500's.
  - Made sure that `BadRequestHttpException`'s are thrown where appropriate in
  `SummaryControllerProvider`.

* Prettifying Test Artifacts

Just pretty printing the test artifacts so that it's easier to see when
something changes.

* Reformatted Code

Just some simple reformatting.

* Removing unused code

We don't have any use cases for the `filters` parameter in this request. If we
do eventually have one, we can come back and snag this code.

* Updated statistics per recent `avg` statistic updates

* Added more descriptive error messages

Needed some additional information to help track down statistic irregularities.

* Update to add date validation to new REST stack

Just added a `checkDateRange` function that will throw a 400 if either date is
not in the correct format or if the start date is after the end date. Also
removed the previously added date validation in the `TimeAggregationUnit` code.

* derrrr

* Adding some documentation to BaseControllerProvider

* Minor rework to date/time validation

Updated the `checkDateRange` function to work like the date validation code that
was previously added to `Query` as the diff'ing was waaaay easier this way.

Also updated the expected output to account for the new error message.

* Minor Cleanup

Removed the unused `duration_change` listener as we no longer base our dates on
the duration toolbar.

Added documentation for the `formatDate` function.

* Updated `getDateTime` -> `getTimestamp`

The `getDateTime` function was no long returning a `DateTime` object but was
instead returning a unix timestamp. The function name / variables created from
this function have been renamed accordingly.

* Removing Unused function argument

The `$tz` argument was for a previous iteration of the code so as it's not being
used it is being removed.

* asdfj