Added GroupByDomain and updated systems accordingly
- Added the new Cloud Realm GroupByDomain class
- The appropriate configuration files were updated.
- New ETLv2 configuration files were added to support the new group by
- Updated some of the integration tests to automatically generate the expected
  test output if not already present.
- Updated the expected test outputs to take into account having the new
  GroupByDomain class.
- Updating the default submission_venue_id to match previous

Adding Documentation and removing unused file

- Just adding documentation to a few classes
- Removed the `.old` test file that shouldn't have been committed

Adding an appended newline for posix compliance

Ensuring that when a new test artifact is created that it is posix compliant (
i.e. that it has a trailing newline ).

Simplifying assertion

The `if/else` was added during testing due to some unexpected errors that were
being encountered. After re-testing I found that it wasn't needed and as such,
is being removed.

Adding default `domain` regression test artifacts

Adding 'Unknown' Domain

Adding an 'Unknown' Domain to utilize when one is not available / identified.