Allow colour override for all datasets
Previously the colour override only worked for the first dataset.
Now the colour override works for all datasets. The datasets that use
the auto colour will get the next colour in the list. The list pointer is
updated when a dateset colour is overridden. This should minimise the chances
of two datasets getting the same colour.