Ensure resource[fact|types] table is pre-populated.
The isRealmEnabled() function is used by the ingestor to determine
which realms should be shredded / ingested. Unfortunately this function
relies on the resourcefact and resource_types tables to be populated.
These tables are populated by the ingestor only if there are realms enabled.
Needless to say this is quite a conundrum. For those interested in the
idea of self-reference, I highly recommend the book Gödel, Escher, Bach:
an Eternal Golden Braid by Douglas Hofstadter.

This bug was not seen in the CI builds because the cloud realm is
hardcoded to run an action that populates the resource* tables outside
of the ingestor code.

This change also reverts the (now unneeded) change to the pipeline run
on the ingestAllStaging path.