Workaround inconsistent return types in XDMoDConfiguration factory
The new configuration API has an odd feature where if the parsed data
has a top level json array and there is one entry in the array then the
entry is returned, If there are multiple entries in the array then an
array is returned. This requires the developer to handle the one and
many cases seperately.

This change adds the one- and many- handling code. I hope that in
future we can remove this and change the confguration API to be
consistent. See https://app.asana.com/0/342819846538629/1146629814310601