Updating cloud event schema and coercing types
During testing we found that the schema validation for cloud events was not
actually validating anything due to a misconfiguration of the schema file.
Instead of failing it just ignored everything.

After updating the schema we also needed to ensure that type coercion was
enabled as the data files encode `2` as `"2"`.