Improve input validation for get_rawdata endpoint.
The get_rawdata metric explorer endpoint had some rather
obfuscuted code that checked for valid data in the start and
limit variables. This code was difficult to understand and
it also triggered false positives in the code analysis software.
There was also a minor inconsistency if the start parameter
was missing from the request. The row data were returned
as if the start parameter was 0 but the totalAvailable metric
was not returned.

This update changes the code to use the standard php filter
API to ensure sanitized input and now has consistent behaviour
for the totalAvailable metric.

Also added tests.