Clear cache operation will remove image files too.
The internal dashboards 'Empty Report Image Cache' only clears the
database contents. The code that retrieves the images checks the
filesystem before the database, so if the file is already on the
filesystem it will served and the database cache will never be
consulted.

This update adds a call to the existing XDReportManager function that cleans up
all report image files.