Adding transactions to `acl-config`
There have been a couple of instances now where `acl-config` has not completed
successfully which has left the db in an inconsistent state. We previously had
limited protection against this for the "non-managed" tables ( user_acls,
user_acl_group_by_parameters). But really we should be protecting the entire
process against this class of error. To that end `acl-config` has been modified
to work as such:
  - we create a backup table for the "non-managed" tables. this will be used
    later in the process.
  - a `_staging` table is created for each acl related table. This new staging
    table is identical to the production table in all but name.
  - we process each module, hierarchy, and set of realms, group bys statistics
    etc. but instead of inserting the information into the production tables we
    populate the staging tables.
  - we populate the staging version of the "non-managed" tables from the backup
    tables.
  - drop the "non-managed" backup tables.
  - In a single DB transaction, we DELETE FROM all of the production tables and
    then INSERT INTO all the data from the corresponding staging tables.
  - finally, we drop the staging tables as they are no longer required.

This new process means that none of the information in the production tables is
modified until we reach the second to last step, and this entire step is
protected by a DB transaction.