Update email notification text to be less mysterious.
The previous email text did not say what had actually changed, what
steps had been taken nor what (if anything) the poor user was supposed
to do about this.

This text conjures up the image of some sort of Orwellian bureaucracy
where sinister faceless operatives were updating your records behind the
scenes and you are allowed to know that the record had been modified
but are not permitted to see it or be told what actually changed.