Fix stderr checkbox disablement in Metric Explorer.
The "Err Bars" and "Err Labels" columns in the "Data" menu in the metric
explorer were incorrectly shown or incorrectly not shown in certian
circumstances.

a) if a chart was 'geared over' from the Usage Tab then the "Data" menu would
always show the "Err Bars" and "Err Labels" checkboxes regardless of whether
they should have been shown.

b) if data were added to the metric explorer from the "Metric Catalog" then
the "Data" menu would never show the checkboxes regardless of whether they
should have been shown.

If data are added to a chart via the "Data Series Definition" dialog then the
checkboxes are correct.

This change fixes it so the checkboxes work correctly.