Don't bother trying to set headers if not possible.
We often get error log reports with messages like the following:
```
PHP Warning:  Cannot modify header information - headers already sent by (output started at /opt/xdmod/share/html/index.php:180) in /opt/xdmod/share/configuration/linker.php on line 179
```

These are of course completely useless and are caused by the error
handler trying to set the headers and failing. The issue is that the
orignal message that the error handler is supposed to output does not
make it to the logs.

This change ensures that the exception handler only tries to set headers
if they have not already been sent and updates the documentation to
encourage folk to look in the xdmod exceptions log.