Enforce realm acls for show raw data endpoints.
Previously, even though the raw data was protected by ACLs the list of
realms was not. So a user could see all realms that supported raw data
access (even if they did not have permission to see the realm itself).

This scenario is not seen in the default open xdmod installs since all
uses have access to jobs and supremm realms. This is seen in the XSEDE
version where different roles have different realm access.

Also add the capabilty to disable a realm from raw data access via the
config file.