Improve search history panel UI.
The search history in the job viewer always displayed all possible
realms that could have search history regardless of whether there were
any saved searches. This was reported as a bug by a user. The realm
nodes in the tree would have a triangle mark indicating that the could
be unfolded, but when the unfold was clicked then it would dissaper.

This change makes it so that the search history only shows the realm
folder nodes for realms that have search history. If there a no saved
searches then an assist panel is shown.