Move lining to githbu action
This moves the linting to a github action. This gives us a very slight
speed improvement due to parallelism and also makes it so that code
linting (which is important) is separated out from the functional build
tests which can speed up development. The unit tests and source code
install are still done in the circle CI upgrade path.