Fix bugs in Configuration classes.
Ensure function prototype is correct for cleanup() function.
Ensure member variables are declared before they are used.

These errors showed up when I was running the unit tests on my ubuntu
box (php 7.2.24). It seems that earlier versions of php let you get away
with it but not in later versions.