Fix multi-timeseries dataset-show-remainder bug.
If you have multiple datasets in the metric explorer and tick 'show
remainder' then the remainder calculation is incorrect for all but the
dataset with the largest series count. This is because the maximum
number of series per dataset is used to compute the remainder rather
than the number of series for each dataset.

Also added a test to confirm the correct behaviour.