Added regression tests for chart filters.
These tests confirm that the chart filters filter the data correctly
and have the correct chart subtitles. All filters available for each
realm are tested with a single metric and group by none.

This also includes the machinery to generate expected test results.

The test for Cloud Filter by instance type is absent. This is because
this filter does not work. Once the instance type filtering is fixed
then a test should be added.