Improve robustness of integration and regression tests.
The integration and regression tests should not get their
test parameters from the system that they are testing. If the
system to test is broken then the test may erroneously pass.

In the case of the current fixes the old code was querying XDMoD
to find out what realms existed and then only testing the realms
that XDMoD reported as present. This is not the way it should be done.
What should actually happen is that the developer sets up XDMoD
with the desired set of realms and then the tests check that
the desired set of realms are actually present.