Remove unnecessary redirect in the job viewer rest stack.
The job viewer rest api has an unusual code path to get the list of jobs
for a given saved search. There is a 302 redirect when the interface
gets the list of jobs. There is no 302 redirect for any of the other
equivalent queries (such as list of information for a job). All of the
required information to respond to the request is in the original request.

This pull request removes the redirect. The path that wasa being
redirected to still remains so as to not break compatibility with any
external systems (they both call the same underlying function).