Fix bug in highchart wrapper code.
The code tries to shade the legend background when the data in the chart
are restricted by roles. This happens e.g. when you have user role and
try to plot some supremm data. Pie charts do not have legends so the
series.legendItem property is absent. This change makes sure the legend
background colour is only changed if there is a legend to change.