GroupBy refactor bugfix.
This change fixes an issue seen with the group by Grant Type in the XSEDE realm.
and is seen when using a group by configuration that has multiple joins and the
dimension table primary key is not stored in the fact tables.

See pull https://github.com/ubccr/xdmod/pull/1288 for more context.