Finally removing Roles / UserRole tables
We have been running for some time without actually utilizing these tables and as such
it should be safe to remove them. These changes will make sure that the tables are not
created / referenced in ETL actions as well as making sure that they are removed in
existing installations when an upgrade is performed.