Updating functions called
`resetUserSettings` is meant to make sure that the user fields `originalValue`
property matches up with the field's current value. Typically this is done after
changes have been saved.

`revertUserSettings` is meant to make sure that the fields current value is
equal to the original value. Which is the desirable behavior in this case.

So while ultimately the form functioned as intended when strictly viewing the
interface, this was more by accident then by design. It also causes confusion
while attempting to read / review the code which is never desirable.