Removing the setting of `startValue`
`startValue` is defined by ExtJs as "The value that the Field had at the time it
was last focused. This is the value that is passed to the change event which is
fired if the value has been changed when the Field is blurred.". And as we do
not have any code that currently operates on the field value passed to its
change event, there's no reason to be updating it here.