Updating Internal Dashboard Existing Email (#1095)
* Updating Internal Dashboard Existing Email

- There was an issue found when updating an existing user's email address.
  Namely, the updated email would not show after saving.
- Added additional UI tests that exercise both changing / discarding & changing
  / saving each of the User Settings ( e-mail, user type, mapped person, and
  institution ).

* Fixing eslint issues

* Removing extraneous function call

This function sets field.value = field.originalValue, since we already assume that field.value
is the up to date value ( hence calling resetDirtyState / resetUserSettings to set
field.originalValue && field.startValue = field.value ). This is an extraneous function
call and can be removed.

* Updating functions called

`resetUserSettings` is meant to make sure that the user fields `originalValue`
property matches up with the field's current value. Typically this is done after
changes have been saved.

`revertUserSettings` is meant to make sure that the fields current value is
equal to the original value. Which is the desirable behavior in this case.

So while ultimately the form functioned as intended when strictly viewing the
interface, this was more by accident then by design. It also causes confusion
while attempting to read / review the code which is never desirable.

* Removing the setting of `startValue`

`startValue` is defined by ExtJs as "The value that the Field had at the time it
was last focused. This is the value that is passed to the change event which is
fired if the value has been changed when the Field is blurred.". And as we do
not have any code that currently operates on the field value passed to its
change event, there's no reason to be updating it here.

Co-authored-by: Ben Plessinger <plessbd@users.noreply.github.com>