Allow the internal dashboard to load even if a user account has an empty name.
The Internal Dashboard -> User Management -> Existing Users tab will sit spinning
in "Loading" if there exists an existing user that has an empty field in any of
the settings (name, email, etc).

This change allows the interface to load so that an admin can actually fix the
problem. Without this change the only way to fix the issue appears to be by directly
modifying the contents of the database.

Note that the missing name issue has been seen on metrics prod and also occurred while
we were testing the XDMoD tutorial. In both cases the undelying cause was a misconfiguration,
(which needed to be rectified) but this change at least allows the resultant problem
to be fixed.