Rename internal dashboard button from "Dashboard" to "Admin Dashboard" (#1390)
Also update the tooltip to be slightly less uninformative.

This increases the width of the button. However there is still plenty
of space in the top toolbar. The width of the page that causes the
Help button to disappear is still much narrower than any usable width
for the rest of XDMoD.