Only shown the New User Tour Dialog once per session.
The new user tour dialog checking-and-showing code runs every time the
dashboard is loaded. The dashboard is reloaded when you navigate to it
and the data have changed. I.e. multiple times per session. This change
sets a flag so that the tour checking code only runs once.

While investigating this issue I noticed another bug where the response
data was being parsed even in the error case (which lead to a unhandled exception
in the javascript). This mitigates this problem by switching the
code to use the success callback function. In the error case the request
will be retried the next time the dashboard loads. I saw the load failure
case when using XDMoD on a noisy wi-fi connection.
Fixes issue https://app.asana.com/0/808093868887967/1140113137533641