Expected Test Updates made to accomodate Centos8 changes
All of these changes are due to the versions of software being updated on
Centos8 vs. Centos7.

- normalized_table_definition.json: Default values are now being returned as
single quoted strings now in MariaDB 10.3.17.

- UsageExplorerTest.php: w/ PHP 7.2 installed these response header values have changed and needed to
be updated.

- RegressionTestHelper.php: w/ PHP 7.2 installed, when dealing with exported CSV
data we sometimes encounter and expect JSON data instead. For instance, when a
user requests data that they do not have access to a JSON object will be
returned. Unfortunately we compare this JSON data as strings and not as objects
/ arrays. This coupled with JSON Pretty Print not including a new line after the
opening square bracket of an empty array ( example below ):

```json
{
    "property1": "",
    "property2": [

    ]
}
```

Versus

```json
{
    "property1": "",
    "property2": []
}
```

This meant that these tests were failing. I've just added a secondary test if
the intitial `$expected === $csvdata` condition fails that tests if $expected /
$csvdata are actually json data.