Adding a new log level conversion function
We previously added a conversion function for going between monolog's log levels
& ccr's levels so this function just provides for going the other way.

These changes were needed after noticing a marked increase in logging in the
Shippable build. After having a think about things it became obvious that we
were providing PSR ( 0-7 ) Log levels to Monolog. As all of PSR's levels are
less than Monolog's DEBUG level ( 100 ), we were always logging at the debug
level. By adding this new conversion function & utilizing it in conjunction w/
the Log Handlers provided by CCR\Log we can ensure that the default log level is
provided correctly to these handlers.

The addition of the `if` statement to `CCR\Logger::log` is for the same reason
as the changes in CCR\Log. It allows the providing of PSR log levels to function
correctly. Specifically this was found while testing the logging in `linker.php`
which directly calls `$logger->log( ... );` as opposed to say `$logger->error(
... );`. And while we tend to use the named log functions in XDMoD, there are a
number of instances of using `log` in XDMoD's modules.