Improve robustness of full groupby checking.
This pull request combines two tightly coupled changes. The underlying issue is
that the ETLv2 code had a deficient check for the correct grouping in aggregate
queries and the etlv1 code autogenerated incorrect grouping configuration.

This change improves the robustness of the ETLv2 groupby check and fixes the
incorrect grouping configuration.

The old ETLv2 code had a php-based check for valid SQL. This didn't handle all
the possible valid cases and allowed an invalid case through. The change
removes this code and instead sets the SQL mode ONLY_FULL_GROUP_BY to get the
SQL server to do the check. Note that not the error is now handled at a
different point in the processing pipeline.

The etlv1 code now uses sql formula in the group by if an sql formula is
defined and the column name in the source table is identical to the column name
in the destination table.