Updating qa-test-setup to include remote branch for Shippable
It was found that while the new qa setup worked well for checking XDMoD
modules ( which have a remote upstream added in shippable.yml ), it had started
failing to pass when run for XDMoD proper due to not being able to identify a
remote branch to check against. Thankfully Shippable provides this information
in the `$BASE_BRANCH` environment variable so, if we're running in Shippable
then update the `build_args` to specify the BASE_BRANCH as the remote for
`build.sh`.

asdkfj