Reverting changes to global_uncaught_exception_handler
These changes were not necessary as it was really the `$logger` specified in
`handle_uncaught_exception` that was in error. Specifically it was that the
original instantiation was a direct call to `\Log::factory` and not
`Log::factory`. The difference being that `\Log::factory` is the PEAR Log
factory and not `\CCR\Log::factory`.