Reconfiguring logger for HostListParserTest
This logger was originally instantiated utilizing PEAR Log's factory method:
```php
$logger = \Log::factory('console', '', 'test', \CCR\Log::DEBUG);
```

This equates to a console logger w/ a log level of DEBUG and an ident of 'test'.

To ensure that we create a logger with the same functionality the way in which
it is configured needed to be changed. Instead of `\CCR\Log::factory` we use
`\CCR\Log::singleton` as that more closely mimics the way PEAR Log functions. We
also needed to explicitely add explicit entries for mail, db and file with false
values so that the resultant logger is only a console logger. The loggers ident
was also changed to `test`.