Add data endpoint that can parse webserver logs and regex ingestor.
This adds a couple of libraries that can parse webserver
log files, parse user agent strings and a geoip lookup library.

The geoip library is Apache Licenced, but the corresponding
geoip database is not redistributable so is not included.

The WebServerLogFile endpoint has configuration settings to
specify the log file format (and hence the records extracted) and
a config setting for the path to the goeip database. If there
is no geoip database then it will still work.

The regular expression ingestor supports using the PHP
preg_filter() function to transform data.

See the xdmod-ood module for CI tests and usage etc.